fikset query

+++ src/main/kotlin/no/nav/k9/tjenester/avdelingsleder/nokkeltall/AlleOppgaverDto.kt
@@ -1,11 +1,11 @@
 package no.nav.k9.tjenester.avdelingsleder.nokkeltall
 
-import no.nav.k9.kodeverk.behandling.BehandlingType
-import no.nav.k9.kodeverk.behandling.FagsakYtelseType
+import no.nav.k9.domene.modell.BehandlingType
+import no.nav.k9.domene.modell.FagsakYtelseType
 
 data class AlleOppgaverDto(
     val fagsakYtelseType: FagsakYtelseType,
     val behandlingType: BehandlingType,
     val tilBehandling: Boolean,
-    val antall: Long
+    val antall: Int
 )
